Add support for scheduled recirculation
This PR adds support for controlling when scheduled recirculation is active, which allows Home Assistant to replicate the weekly scheduling capability of the NaviLink app. It has been tested against an NPE-240A2 water heater, but should hopefully work on other models as well.

It adds two new control packets `SCHEDULED_RECIRC_ON_CMD` and `SCHEDULED_RECIRC_OFF_CMD`. The first one tells the Navien unit that it can recirculate when it sees fit (and also triggers an immediate recirculation cycle), and is what gets sent by NaviLink when either a scheduled recirculation time block begins, or when you turn off the weekly schedule (which puts the unit back into "always recirculate" mode). And the second is sent by NaviLink when a recirculation time block ends.

At a low level, whether or not recirculation is enabled is represented by bit 1 of byte 33 (`recirculation_enabled`) being set, and sending these commands causes that bit to be set or cleared. The commands will set or clear that bit regardless of whether the unit is in "HotButton" mode or one of the three scheduled modes ("Always", "Intelligent", or "Weekly"); however if the unit is in "HotButton" mode, no actual recirculation will occur, at least on an NPE-240A2. When the unit is actively recirculating, this is reflected by `heating_mode` changing to 0x08 (`HEATING_MODE_DOMESTIC_HOT_WATER_RECIRCULATING`).

From Home Assitant's perspective, the sending of these packets is controlled by a new switch, "Allow Recirculation". Two new sensors are also added by this PR:

* "Scheduled Recirculation" is a binary sensor that will show "Running" if a recirculation cycle is active (it reflects whether the "Heating Mode" sensor is "DHW Recirculating"; I had coded it up before I had merged in the changes that added the "Heating Mode" sensor, so it is a little bit redundant)
* "Hot Button Recirculation" is a binary sensor that will show "Running" if a HotButton-triggered recirculation cycle is active (which is indicated by bit 0 of `recirculation_enabled` being set. For whatever reason, a HotButton-triggered recirculation cycle just shows in `heating_mode` as "Domestic Hot Water", the same as if a tap was running.

In addition, I changed the "Recirculation Status" binary sensor to a text sensor named "Recirculation Mode", which will have the value `HotButton` or `Scheduled`, as that more accurately reflects the meaning of bit 1 of `system_status`.

Regarding the two "recirculation running" sensors, I debated whether to combine them into a single sensor, since they can't both ever be true at the same time, and the "scheduled recirculation" one by itself is a bit redundant with the recently-added "Heating Mode" sensor. I'm open to suggestions.

Last but not least, this PR includes a Home Assistant automation blueprint that allows the user to set up a recirculation schedule for their water heater, using a Schedule helper. It also allows for recirculation to be triggered by other switches or buttons (and for buttons, it will basically simulate the HotButton functionality by enabling recirculation until a single cycle completes).